# [xy] Bump up pandas version to improve perf.