--- a/requirements.txt
+++ b/requirements.txt
@@ -11,7 +11,7 @@ itsdangerous==2.1.2
 joblib>=1.1.0
 numpy~=1.21.0
 numpyencoder==0.3.0
-pandas==1.1.3
+pandas~=1.4.0
 pyarrow==8.0.0
 python-dateutil==2.8.2
 pytz==2022.1