# Summary
<!-- Brief summary of what your code does -->
Bump up pandas version to improve perf. pandas [1.4.0](https://github.com/pandas-dev/pandas/releases/tag/v1.4.0), [1.3.0](https://github.com/pandas-dev/pandas/releases/tag/v1.3.0) and [1.2.0](https://github.com/pandas-dev/pandas/releases/tag/v1.2.0) all have perf improvements.

The old pandas version was slow on computing nunique (it took 5 seconds to calculate nunique for 100k unique rows). It's instant using the new pandas version.

Performance comparison (a dataset with 100k rows and 32 columns.
Old pandas version
* infer_column_types took 10s.
* calculate_statistics took 15s

New pandas version
* infer_column_types took 5s.
* calculate_statistics took 1.1s

# Tests
<!-- How did you test your change? -->
tested locally

cc:
<!-- Optionally mention someone to let them know about this pull request -->